feature(ui-ux): mock api for latest data
